Backend/requirements.txt
annotated-types      0.7.0
anyio                4.6.0
blinker              1.8.2
Brotli               1.0.9
certifi              2024.8.30
charset-normalizer   3.4.0
click                8.1.7
colorama             0.4.6
contourpy            1.3.0
cycler               0.12.1
dnspython            2.7.0
email_validator      2.2.0
fastapi              0.115.3
fastapi-cli          0.0.5
filelock             3.16.1
Flask                3.0.3
Flask-Cors           5.0.0
fonttools            4.54.1
fsspec               2024.10.0
h11                  0.14.0
httpcore             1.0.6
httptools            0.6.1
httpx                0.27.2
huggingface-hub      0.26.2
idna                 3.10
itsdangerous         2.2.0
Jinja2               3.1.4
kiwisolver           1.4.7
markdown-it-py       3.0.0
MarkupSafe           3.0.1
mdurl                0.1.2
mkl_fft              1.3.10
mkl_random           1.2.7
mkl-service          2.4.0
mpmath               1.3.0
networkx             3.4.2
numpy                2.1.2
orjson               3.10.10
packaging            24.1
pandas               2.2.3
pillow               11.0.0
pip                  24.3.1
pydantic             2.9.2
pydantic_core        2.23.4
pydantic-extra-types 2.9.0
pydantic-settings    2.6.0
Pygments             2.18.0
pyparsing            3.2.0
PySocks              1.7.1
python-dateutil      2.9.0.post0
python-dotenv        1.0.1
python-multipart     0.0.12
pytz                 2024.2
PyYAML               6.0.2
regex                2024.9.11
requests             2.32.3
rich                 13.9.2
safetensors          0.4.5
setuptools           75.3.0
shellingham          1.5.4
six                  1.16.0
sniffio              1.3.1
starlette            0.41.0
sympy                1.13.1
tokenizers           0.20.1
torch                2.5.0
torchaudio           2.5.0
torchvision          0.20.0
tqdm                 4.66.6
typer                0.12.5
typing_extensions    4.12.2
tzdata               2024.2
ujson                5.10.0
urllib3              2.2.3
uvicorn              0.32.0
watchfiles           0.24.0
websockets           13.1
Werkzeug             3.0.6
wheel                0.44.0
win-inet-pton        1.1.0